Refactored ROS driver, improved perfromance (Callback 2ms), correct distance calculation now included
