--- include/nimbus-ros/driver.hpp
@@ -1,1 +1,0 @@
-
