--- src/driver.cpp
@@ -1,3 +1,0 @@
-#include "nimbus-ros/driver.hpp"
-
-
